Anime-B-Gone
The world is a better place without demihumans. Catgirls and Foxgirls do not fit the vision for the server and tend to attract a sort of playerbase we do not wish to have - it was generally agreed upon a few months ago that we were better off without them, so this is me finally getting around to it.