Variable defining, attempted to figure out the spriting for stock and no-stock versions
I'm over my head. I can't get the cycler to work - I think I need to bind it to a different key, or perhaps the regular Cycler just... doesn't work on shiptest.

I'm going to test the regular Cycler, and if it doesn't, then I'll just try letting the TB83 use 3 rounds of shotgun and/or rifle instead of a set 2 shot 1 rifle... and if it does, well, I've just got to figure out how to make mine do so as well, don't I?